try cluster-level optimize and pipeline

+++ cluster_hnsw_nsg/tests/combination_test/cluster_knn_knn_search.cpp
@@ -285,7 +285,7 @@ int main(int argc, char** argv) {
         // 选择前k个结果
         std::vector<unsigned> final_results;
         final_results.reserve(k);
-        for (int m = 0; m < k && m < all_results.size(); m++) {
+        for (int m = 0; m < k && m < (int)all_results.size(); m++) {
             final_results.push_back(all_results[m].second);
         }
